handle dashes in vpc names
Switch to using database names (the id without env name prefix) for most functions.
Rename database_class to database_name everywhere.
Its not possible to extract the database name from a database id
without knowing the environment of the database so a number of functions
now need the env_name as an argument