Make alarms work with dynamic ASG names
DiscoAlarms was assuming that the names of autoscaling group was static,
but that is no longer true due to changes necessary for blue/green
deployment. Instead, when alarms are being configured for a hostclass,
the autoscaling group name can now be provided along with that call, or
the DiscoAlarmsConfig now has the capability to grab the autoscaling
group name if one was not provided and the namespace is 'AWS/EC2'.

Additionally, because autoscaling group names are not part of the
namespacing of the alarm names, there was a possible problem where
blue/green deployment might update alarms to point to the new
autoscaling group, destroy that autoscaling group when tests fail, and
then not repoint the alarms to the old autoscaling group. To resolve
this, alarms are not created for autoscaling groups that are marked
with 'is_testing'. Instead, alarms will be created for that autoscaling
group after it passes tests, similar to how we handle ELB switchover.

Also added some unit tests to ensure that we try to lookup the
autoscaling group name when it isn't provided.