Fix blue/green not respecting the pipeline
In the case of their being no ASG, blue/green deployment would always
default to '1' for the sizing of the new ASG, even if there was an entry
in the pipeline file. This change should fix the broken if condition
that lead to this state and make sure that the pipeline entry is
respected.