Handle spinup errors in blue/green deployments
As has been demonstrated, sometimes AWS cannot provide enough instances
for an ASG and this causes the spinup function to fail and throw an
exception. Now, we should catch any exceptions from the spinup command
and try to teardown the newly created testing ASG (and ELB if necessary)
so that the ASGs are not left in a bad state.

Added a number of unit tests to try to cover the different combinations
of already existing old ASGs and newly created ASGs to make sure that we
don't accidently destroy an ASG that is already in use can cause
downtime.