Increase ELB Instance State timeout
It looks like fairly frequently instances are not able to attach to an
ELB within the current timeout. This might be due to delays in AWS
making instances routable. AWS doesn't appear to give any guidelines on
how long an instance should take to attach to an ELB, so trying a nice
round number like 10 minutes.