Fix timed autoscaling for disco_deploy.py
This is a bit of a refactor to make disco_deploy.py understand and
respect the timed autoscaling format we use in the pipeline file. Now,
classic and blue_green deployment strategies will remove the existing
timed autoscaling actions and reapply any that were defined in the
pipeline file. One notable is for blue_green deployment: if the
hostclass isn't deployable, then blue_green won't bother to apply the
scheduled actions. The same happens for classic deployment and
non-deployable if the hostclass is not already deployed - the timed
autoscaling actions are not applied. Additionally, hostclasses that take
advantage of timed autoscaling are now expected to respect the sizing
of their already deployed ASG, if one exists.

This required a bit of refactor for the method signatures and structure
of the `handle_*` methods. Now all of them accept the preexisting ASG as
an optional argument, as well as the original pipeline definition if one
exists. The logic for creating the ASG sizing is now centralized in the
two `_generate_*` methods to make everything a bit more DRY. Instead of
passing a pipeline and relying on the logic in provision and spinup to
intepret the current sizing and the timed autoscaling actions, that
logic is now split up. Instead, disco_deploy will use the new
`_generate_*` functions to determine the appropriate sizing, and then
use the new `_create_scheduled_actions` function to create the timed
autoscaling actions as per the original pipeline.

Also fixed a bug where blue_green deployment would break if someone set
`0` as the desired_size in their pipeline file.

Also included some general cleanup of the disco deploy unit tests.